requirements.txt
icalendar>=5.0.2
requests>=2.28.1
argparse>=1.4.0
PyYAML>=6.0
discord.py>=2.1.0 
